requirements.txt
streamlit==1.7.0
requests==2.26.0
spacy==3.1.3
nltk==3.6.3
scikit-learn==0.24.2
matplotlib==3.4.3
wordcloud==1.8.1
python-dotenv==0.19.1
